update enum name
